Fix Issue #2
Added new direct_config function that accepts an object containing
arguments for direct. This closes Issue #2.

Also added a 'call before' argument to direct, which gets called with
an object containing the method, args and callback. I'm not sure how
useful this will prove to be, but it was easy to add and might come in
handy.